requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
httpx
python-dotenv